requirements.txt
click==8.0.3
Flask==2.2.2
email_validator==1.1.3
Flask-Mail==0.9.1
Flask-WTF==0.15.1
gunicorn==20.1.0
importlib-metadata==6.0.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.2
python-dotenv==0.21.1
Werkzeug==2.2.2
zipp==3.11.0
